+++ test/integration/render/run_render_tests.ts
@@ -44,7 +44,7 @@ type TestData = {
     collisionDebug: boolean;
     localIdeographFontFamily: string;
     crossSourceCollisions: boolean;
-    disableTerrainVerticalExtensions: boolean;
+    disableTerrainSkirts: boolean;
     operations: any[];
     queryGeometry: PointLike;
     queryOptions: any;
@@ -736,7 +736,7 @@ async function getImageFromStyle(styleForTest: StyleWithTestData, page: Page): P
                 fadeDuration: options.fadeDuration || 0,
                 localIdeographFontFamily: options.localIdeographFontFamily || false as any,
                 crossSourceCollisions: typeof options.crossSourceCollisions === 'undefined' ? true : options.crossSourceCollisions,
-                disableTerrainVerticalExtensions: options.disableTerrainVerticalExtensions,
+                disableTerrainSkirts: options.disableTerrainSkirts,
                 maxCanvasSize: [8192, 8192]
             });